starting to implement new perlin noise algorithm
