+++ MultiHex/generator/fill_land.py
@@ -1,6 +1,7 @@
 from MultiHex.core import Hexmap, save_map, load_map
 from MultiHex.map_types.overland import *
 from MultiHex.generator.util import *
+from MultiHex.generator.noise import perlinize
 
 from numpy import arccos 
 from numpy import histogram
