+++ MultiHex/generator/full_chain.py
@@ -18,9 +18,9 @@ both refer to the file `config.json` for simulation parameters
 
 def ridge_onward( size, name ):
     fill_land.generate( size, name )
-    sim_weather.generate( size, name )
-    draw_rivers.generate(size, name)
-    biome_maker.generate( size, name )
+#    sim_weather.generate( size, name )
+#    draw_rivers.generate(size, name)
+#    biome_maker.generate( size, name )
 
 def full_sim( size, name ):
     make_ridges.generate( size, name )
