+++ MultiHex/generator/util.py
@@ -11,187 +11,6 @@ from numpy.random import rand
 #                  Prepare Utilities
 # =====================================================
 
-def _is_valid_texture(obj):
-    """
-    Checks the object provided to verify that it is a properly formated texture list
-    """
-    if not (isinstance(obj, list) or isinstance(obj, ndarray)):
-        print("Not List, {}".format(type(obj)))
-        return(False)
-    if not (isinstance(obj[0], list) or isinstance(obj[0],ndarray)):
-        print("Not List, {}".format(type(obj[0])))
-        return(False)
-    if not (len(obj)==len(obj[0])):
-        print("Not rectangular")
-        return(False)
-    if not (isinstance(obj[0][0], float) or isinstance(obj[0][0], int)):
-        print("Not number")
-        return(False)
-    return(True)
-
-def _save_texture(obj):
-    """
-    Saves the texture to a numpy file and also ensures it is a properly formated texture list
-    """
-    assert(_is_valid_texture(obj))
-    where = os.path.join(os.path.dirname(__file__),'.texture.npy')
-    save( where, obj)
-
-def _load_texture():
-    """
-    Loads the saved texture and verifies that it is valid before returning it
-    """
-    where = os.path.join(os.path.dirname(__file__),'.texture.npy')
-    temp = load(where)
-    assert(_is_valid_texture(temp))
-    return( temp )
-
-def sample_noise(x_pos, y_pos, xsize=None, ysize=None, texture=None):
-    """
-    Samples from a texture list-object by interpolating between points in its 2D array. This is done by treating the two indices of the texture as (x,y) coordinates.
-    When a point between those integer indices is requested, the average of the values corresponding to the indices that are floor and roof of that point are used.
-
-    When an xsize and ysize are specified, we stretch the provided x_pos and y_pos to fit within the texture's bounds by treating the x/ysize as the maximum index of the texture. 
-
-    params
-    x_pos   - float, the x position at which to sample
-    y_pos   - float, the y position at which to sample
-
-    """
-
-    # optionally, the user can pass the texture itself to minimize IO
-    if texture is None:
-        try:
-            texture = _load_texture()
-        except IOError:
-            print("File not found while sampling. Generating...")
-            _generate_perlin_texture()
-            texture = _load_texture()
-    else:
-        if not _is_valid_texture(texture):
-            raise ValueError("Something was wrong with the specified texture")
-
-    # we need to stretch the texture to fit the map
-    if xsize is None:
-        xscale_factor = 1.
-    else:
-        assert( isinstance(xsize,float) or isinstance(xsize,int))
-        xscale_factor = float(len(texture))/xsize
-    if ysize is None:
-        yscale_factor = 1.
-    else:
-        yscale_factor = float(len(texture[0]))/ysize
-
-    # get the lower corner in the texture that the requested position is closest to
-    xsam = int(x_pos*xscale_factor)
-    ysam = int(y_pos*yscale_factor)
-
-    # we need to be sure we're sampling within the bounds of the texture
-    if not (xsam < len(texture) and xsam>=0):
-        raise ValueError("Position {} is beyond the extent {}".format(xsam, len(texture)))
-    if not (ysam < len(texture[0]) and ysam>=0):
-        raise ValueError("Position {} is beyond the extent {}".format(ysam, len(texture[0])))
-
-
-    # now just average the pixels around the sampled point
-    value = texture[xsam][ysam] + texture[(xsam+1)%len(texture)][ysam] + texture[xsam][(ysam+1)%len(texture[0])] + texture[(xsam+1)%len(texture)][(ysam+1)%len(texture[0])]
-    return( 0.25*value)
-
-def _generate_perlin_texture(size = 512):
-    """
-    This generates a texture from which to sample in the perlinize function. 
-    """
-    assert(isinstance(size, int))
-    assert( size % 2 == 0 )
-
-    # we need a 2D array of points which will represent our noise texture
-    #    later, we will sample from this texture by interpolating between points on it
-    grid_full = [ [rand() for j in range(size)] for i in range(size) ]
-
-    # now we need some down-sampled maps to apply large-scale fluctuations
-
-    # So first we find the maximum number of downsamples we can perform
-    levels = 1
-    while (size % (2**levels) ==0):
-        levels += 1
-    levels-=1
-    print("Doing {} levels of down-sampling".format(levels))
-
-    downsamples = {}
-    for level in range(levels):
-        eff_size = int(size/(2**(level+1)))
-        downsamples[ level ] = [ [rand() for j in range(eff_size)] for i in range( eff_size ) ]
-
-
-    # now we go add all the layers together
-    for i in range(size):
-        for j in range(size):
-            for level in range(levels):
-                eff_size = int(size/(2**(level+1)))
-                eff_i = int(i/(2**(level+1))) 
-                eff_j =  int(j/(2**(level+1)))
-                grid_full[i][j] += 0.6*((level+1)**1)*downsamples[level][ eff_i ][ eff_j ]
-
-                grid_full[i][j] += 0.1*((level+1)**1)*( downsamples[level][ (eff_i+1)%eff_size ][ eff_j ] \
-                                + downsamples[level][ eff_i   ][ (eff_j+1)%eff_size ] \
-                                + downsamples[level][ (eff_i-1)%eff_size ][ eff_j ] + downsamples[level][ eff_i ][ (eff_j-1)%eff_size ]  )
-
-    maax = max(grid_full)
-    miin = min(grid_full)
-    for i in range(size):
-        for j in range(size):
-            grid_full[i][j] = 2*((grid_full[i][j]- miin)/(maax - miin)) - 1
-
-    _save_texture(grid_full)
-
-
-
-def perlinize( which = os.path.join(os.path.dirname(__file__),'..','saves','generated.hexmap'), attr='_altitude_base', magnitude = 0.20 ):
-    """
-    Injets perline noise into the specified map's attribute. 
-
-    @param which - which map. Either a string specifying the filepath or a pointer to the Hexmap
-    @param attr  - which hex attribute to modify 
-    @param magnitude - scale of noise to inject. 0.05-0.40
-    """
-    
-    # this opens the map / loads the map to the proper variable 
-    if isinstance(which,str):
-        main_map = load_map(which)
-    #elif isinstance(which, Hexmap):
-    #    main_map = which
-    else:
-        raise TypeError("Unexpected type for arg 'which': {}. Expected {} or {}".format(type(which), str, Hexmap))
-
-
-    # should ensure that the attribute given and the magnitude are appropriately typed
-    if not isinstance( attr, str ):
-        raise TypeError("Unexpected type {} for arg 'str', expected {}".format(type(attr), str))
-    if not isinstance(magnitude, float):
-        raise TypeError("Expected {} for 'magnitude', got {}".format(float, type(magnitude)))
-    
-    # We need to make sure the attribute exist in our Hexes! This way we don't waste time creating a texture we don't need 
-    for HexID in main_map.catalogue:
-        if hasattr( main_map.catalogue[HexID], attr):
-            break
-        else:
-            raise KeyError("HexMap does not have attribute {}".format(attr))
-
-    # We need to create a noise texture file
-    _generate_perlin_texture()
-    # then grab it into an object to minimize IO later on
-    texture = _load_texture()
-    for HexID in main_map.catalogue:
-        this_hex = main_map.catalogue[HexID]
-        cent = this_hex.center
-        # use the Hexes location to skew the specified attribute of the hex 
-        new_value = getattr(this_hex, attr) + magnitude*sample_noise(cent.x, cent.y, main_map.dimensions[0], main_map.dimensions[1], texture)
-        setattr( this_hex, attr, new_value)
-        this_hex.rescale_color()
-    save_map( main_map, which )
-
-
-
 def angle_difference( theta_1, theta_2 ):
     """
     Returns the absolute difference between two angles
